add sieve of eratosthenes algorithm

+++ src/main/java/org/karina/calculator/PrimeNumberCalculator.java
@@ -3,5 +3,5 @@ package org.karina.calculator;
 import java.util.List;
 
 public interface PrimeNumberCalculator {
-    List<Long> calculate(long n);
+    List<Integer> calculate(int n);
 }
